feat: Update meta properties and add Open Graph tags for receipt generator
The code changes in the index.html file update the meta properties and add Open Graph tags to improve the SEO and social media sharing of the receipt generator. This includes updating the title, description, keywords, author, and robots meta tags, as well as adding Open Graph tags for title, description, type, and image. These changes ensure that the receipt generator is properly represented when shared on social media platforms.